refactor: extract shared code
