+++ jest.setup.js
@@ -1,0 +1,2 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
+require('dotenv').config();
